Move _specialize_per_device from recurrent to recurrent_internals
